Add new treeview widget and derivatives
